refactor: Converting to camelCase
